# src/docs/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SellMate API",
    "version": "1.0.0",
    "description": "\n# SellMate Platform - Social Commerce Automation Backend API\n\nSellMate is an intelligent social commerce automation platform designed specifically for Nigerian micro-entrepreneurs who sell through Instagram and WhatsApp.\n\n## Features\n- **Unified Conversations**: Consolidate WhatsApp and Instagram messages in one place\n- **AI-Powered Order Detection**: Automatically detect orders from conversations\n- **Profit Analytics**: Real-time business metrics and profit calculations\n- **Customer Management**: Track customer interactions across platforms\n\n## Authentication\nMost endpoints require JWT authentication. Include the token in the Authorization header:\n```\nAuthorization: Bearer <your_jwt_token>\n```\n\n## Rate Limiting\nAPI requests are rate-limited to prevent abuse. Current limits:\n- General: 100 requests per 15 minutes\n- Authentication: 5 requests per 15 minutes\n- AI endpoints: 20 requests per minute\n      ",
    "contact": {
      "name": "SellMate Team",
      "email": "support@sellmate.ng"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.sellmate.ng",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token obtained from /auth/verify-otp or /auth/refresh"
      }
    },
...
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/subscription": {
      "get": {
        "summary": "Get current subscription details",
        "description": "Returns the user's current subscription plan, status, and features",
        "tags": [
          "Subscription"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Subscription details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Subscription"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/subscription/usage": {
      "get": {
        "summary": "Get current usage statistics",
        "description": "Returns usage metrics against plan limits",
        "tags": [
          "Subscription"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Usage statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "orders": {
                          "type": "object",
                          "properties": {
                            "used": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "percentage": {
                              "type": "number"
                            }
                          }
                        },
                        "conversations": {
                          "type": "object",
                          "properties": {
                            "used": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "percentage": {
                              "type": "number"
                            }
                          }
                        },
                        "aiRequests": {
                          "type": "object",
                          "properties": {
                            "used": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "percentage": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/subscription/billing": {
      "get": {
        "summary": "Get billing cycle information",
        "description": "Returns current billing cycle details and next billing date",
        "tags": [
          "Subscription"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Billing information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "currentPeriodStart": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "currentPeriodEnd": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "nextBillingDate": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "amount": {
                          "type": "number"
                        },
                        "currency": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/subscription/payments": {
      "get": {
        "summary": "Get payment history",
        "description": "Returns list of past payments and invoices",
        "tags": [
          "Subscription"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "amount": {
                            "type": "number"
                          },
                          "currency": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "successful",
                              "failed",
                              "pending",
                              "refunded"
                            ]
                          },
                          "paidAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "description": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/subscription/upgrade": {
      "post": {
        "summary": "Upgrade subscription plan",
        "description": "Upgrades the user's subscription to a higher tier plan",
        "tags": [
          "Subscription"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "planId"
                ],
                "properties": {
                  "planId": {
                    "type": "string",
                    "enum": [
                      "starter",
                      "professional",
                      "business"
                    ],
                    "description": "Target plan to upgrade to"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription upgraded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Subscription"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid plan or cannot upgrade"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/subscription/downgrade": {
      "post": {
        "summary": "Downgrade subscription plan",
        "description": "Downgrades the user's subscription to a lower tier plan (effective at end of billing cycle)",
        "tags": [
          "Subscription"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "planId"
                ],
                "properties": {
                  "planId": {
                    "type": "string",
                    "enum": [
                      "starter",
                      "professional",
                      "business"
                    ],
                    "description": "Target plan to downgrade to"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Downgrade scheduled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "effectiveDate": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid plan or cannot downgrade"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}